requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
pyjwt==2.7.0
pytest==8.3.5
pytest-flask==1.3.0
python-dateutil>=2.8.0,<3.0.0
python-dotenv==1.0.1
sqlalchemy==2.0.40
psycopg2-binary==2.9.10
alembic==1.15.2
